+++ playground/app/src/app/Library/Layout/Components/MenuCategory/Component.ts
@@ -9,5 +9,5 @@ export class MenuCategoryComponent {
   public readonly type = "category";
 
   @Input("name") public name?: string = "";
-  @Input("showLabel") public showLabel?: boolean = false;
+  @Input("showLabel") public showLabel?: boolean = true;
 }
