+++ playground/app/src/app/Modules/TextEditor/index.ts
@@ -1,1 +1,11 @@
-export * from "./Module";
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+
+import { TextEditorComponent } from "./Component";
+import { TextEditorRoutingModule } from "./Routing";
+
+@NgModule({
+  declarations: [TextEditorComponent],
+  imports: [CommonModule, TextEditorRoutingModule]
+})
+export class TextEditorModule {}
